Move CasterBoardLayout class into its own file
We're going to extend that class, so let's give it its own file.
